Refactor fetching showlist to Contentful client.
