+++ contentful/graphql/showlistPage.graphql
@@ -7,27 +7,6 @@ query showlistPage($showlistId: String) {
       heroImage {
         url
       }
-      showsCollection(limit: 200) {
-        items {
-          name
-          start
-          end
-          description
-          picture {
-            title
-            description
-            contentType
-            fileName
-            size
-            url
-            width
-            height
-          }
-          hosts
-          producer
-          color
-        }
-      }
     }
   }
 }
